+++ processes/validation.py
@@ -3,7 +3,7 @@ import logging
 import re
 from typing import Union, Set, Tuple
 
-from processes.date_logic import str_to_date_dict, date_dict_to_str, datetime_to_str
+from processes.datetime_logic import str_to_date_dict, date_dict_to_str, datetime_to_str
 
 
 class ValidationError(Exception):
